requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-phone-field==1.8.1
django-phonenumber-field==8.0.0
idna==3.7
mercadopago==2.2.3
mysql==0.0.3
mysqlclient==2.2.4
phonenumbers==8.13.4
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
gunicorn==20.1.0
